Update to use renderer (#3)
It turns out that on 4k displays, updating every pixel on a
full-resolution surface is crazy expensive and slow. Surfaces are
apparently the "old" way to do things anyway.

So now I am using a renderer and a texture, which is just the size of
the original grid, but gets scaled up to the display size.