Remove useless restrict keyword
I think I went a little crazy when I first learned about the restrict
keyword, and used it all over. Now that I have a better understanding of
what it's for, I think I don't really need it (and some usage was just
wrong). This is after analyzing the compiler output a bit.